flutter restructure
