+++ src/pages/integration-builder/builder/frameworks/flutter/filenames.ts
@@ -1,0 +1,5 @@
+export const FILENAME_MAINACTIVITY = "frameworks/flutter/main.dart";
+
+export default function getFileNames(filenames) {
+  filenames.push(FILENAME_MAINACTIVITY);
+}
