+++ src/pages/integration-builder/builder/frameworks/ios/stepContent/register-app.mdx
